refactor: organize component folders
