+++ examples/template-svelte/packages/app/src/components/graphs/graph.svelte
