+++ examples/template-svelte/packages/app/src/components/graphs/legend.svelte
