+++ examples/template-svelte/packages/app/src/components/graphs/selectable-graph-vm.ts
