+++ examples/template-svelte/packages/app/src/components/inputs/slider.svelte
